feat: add status column to traces table
