+++ app-server/src/traces/utils.rs
@@ -1,3 +1,4 @@
+use lazy_static::lazy_static;
 use std::sync::Arc;
 
 use backoff::ExponentialBackoffBuilder;
@@ -9,6 +10,7 @@ use crate::{
     cache::Cache,
     db::{
         self, DB,
+        events::Event,
         labels::LabelSource,
         spans::{Span, SpanType},
         trace,
@@ -31,6 +33,11 @@ pub fn json_value_to_string(v: &Value) -> String {
             .join(", "),
         _ => v.to_string(),
     }
+}
+
+lazy_static! {
+    static ref EXCEPTION_REGEX: Regex =
+        Regex::new(r"^(exception|exception\.(type|message|stacktrace))$").unwrap();
 }
 
 /// Calculate usage for both default and LLM spans
@@ -89,6 +96,7 @@ pub async fn record_span_to_db(
     span_usage: &SpanUsage,
     project_id: &Uuid,
     span: &mut Span,
+    events: &Vec<Event>,
 ) -> anyhow::Result<()> {
     let mut trace_attributes = TraceAttributes::new(span.trace_id);
 
@@ -96,6 +104,13 @@ pub async fn record_span_to_db(
     trace_attributes.update_end_time(span.end_time);
 
     let mut span_attributes = span.get_attributes();
+
+    events.iter().for_each(|event| {
+        // Check if it's an exception event
+        if event.name == "exception" || EXCEPTION_REGEX.is_match(&event.name) {
+            trace_attributes.set_status("error".to_string());
+        }
+    });
 
     trace_attributes.update_session_id(span_attributes.session_id());
     trace_attributes.update_trace_type(span_attributes.trace_type());
